+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.1.3
-- Bump Flutter SDK range to ">=2.18.0 <4.0.0".
+- Bump Flutter SDK range to ">=2.17.6 <4.0.0".
 
 ## 0.1.2
 - Add support for generating enum extensions.